requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-dotenv
requests
requests[socks]
python-nmap
dnspython
slowapi
apscheduler>=3.10
shodan
PySocks
psycopg2-binary
passlib[bcrypt]
bcrypt==4.0.1
email-validator
psutil
PyMuPDF
python-multipart
beautifulsoup4
